packages/keyvault-llm/keyvault_llm-0.1.4.dev0.tar.gz/keyvault_llm-0.1.4.dev0/src/keyvault_llm/tests/test_client.py
import unittest
from unittest.mock import patch
from keyvault_llm.client import KeyVaultClient

class TestKeyVaultClient(unittest.TestCase):

    @patch('keyvault.client.requests.get')
    def test_get_key(self, mock_get):
        mock_get.return_value.json.return_value = {'test_key': 'test_value'}
        mock_get.return_value.status_code = 200

        client = KeyVaultClient()
        result = client.get_key('test_key')

        self.assertEqual(result, 'test_value')
        mock_get.assert_called_once_with('http://localhost:38680/get_key/test_key')

    @patch('keyvault.client.requests.get')
    def test_list_keys(self, mock_get):
        mock_get.return_value.json.return_value = {'keys': ['key1', 'key2']}
        mock_get.return_value.status_code = 200

        client = KeyVaultClient()
        result = client.list_keys()

        self.assertEqual(result, ['key1', 'key2'])
        mock_get.assert_called_once_with('http://localhost:38680/list_keys')

if __name__ == '__main__':
    unittest.main()
